School Lunch Menu Ordering System
Code so far has Login, and Registration for existing and new customers. It has link to file which has menu items, and menu items can easily be changed and upgraded. Quantity can be selected. Code still requires users to be able to select multiple menu items and link to payment.